Comment out clipboard listening in lifecycle methods
This commit disables the calls to `_clipboardService.StopListening()` in the `OnSleep` method and `_clipboardService.StartListening()` in the `OnResume` method. As a result, the application will continue to listen to clipboard events when it goes to the background and when it resumes, potentially affecting clipboard data interactions during these lifecycle events.